refactor:useQuery, store코드 entities로 이동
