+++ src/entities/goal/model/index.ts
@@ -1,2 +1,4 @@
 export * from './queryKeys';
 export * from './types';
+export * from './useGetGoal';
+export * from './useGoalStore';
