+++ src/entities/study/model/index.ts
@@ -1,3 +1,5 @@
 export * from './queryKeys';
 export * from './types';
 export * from './useStudyGroup';
+export * from './useGetStudy';
+export * from './useStudyStore';
