+++ src/entities/study/model/useGetStudy.ts
