--- src/features/get-goal-list/model/index.ts
@@ -1,2 +1,0 @@
-export * from './useGetGoal';
-export * from './useGoalStore';
