--- src/features/get-study-list/model/index.ts
@@ -1,2 +1,0 @@
-export * from './useGetStudy';
-export * from './useStudyStore';
